fixed circular imports in userdataschema.py, move UserDataSchemaProvider to schemaprovider.py
